ndap/requirements.txt
fastapi
pandas
kafka-python
requests
clickhouse-connect
influxdb-client
python-dotenv
scapy
clickhouse-driver
scikit-learn
python-multipart
uvicorn
tabulate
imbalanced-learn
xgboost